+++ src/geometry/lines.rs
@@ -16,37 +16,37 @@ pub type OutCode = i8;
 
 /// Represents a 2-dimensional line of finite length. The line is defined by a start point and an end point.
 pub struct Line2 {
-    pub points: [Point2<f64>; 2]
+    pub points: [Point2<f32>; 2]
 }
 
 impl Line2 {
-    pub fn new(start: Point2<f64>, end: Point2<f64>) -> Line2 {
+    pub fn new(start: Point2<f32>, end: Point2<f32>) -> Line2 {
         Line2 {points: [start, end]}
     }
 
     /// Returns the start point of the line.
-    pub fn start(&self) -> Point2<f64> {
+    pub fn start(&self) -> Point2<f32> {
         self[0]
     }
 
     /// Returns the end point of the line.
-    pub fn end(&self) -> Point2<f64> {
+    pub fn end(&self) -> Point2<f32> {
         self[1]
     }
 
     /// Returns the length of the line.
-    pub fn length(&self) -> f64 {
+    pub fn length(&self) -> f32 {
         (self[1] - self[0]).norm()
     }
 
     /// Scales the line by a constant. Both the start point and the end point are multiplied by the scalar value.
-    pub fn scale(&self, scale: f64) -> Line2 {
+    pub fn scale(&self, scale: f32) -> Line2 {
         Line2::new(self.points[0] * scale, self.points[1] * scale)
     }
         
     // Cohen–Sutherland clipping algorithm clips a line from against a rectangle with 
     // diagonal from (min_x, min_y) to (max_x, max_y).
-    pub fn clip(&self, min_x: f64, max_x: f64, min_y: f64, max_y: f64) -> Option<Line2> {
+    pub fn clip(&self, min_x: f32, max_x: f32, min_y: f32, max_y: f32) -> Option<Line2> {
 
         let left = 1;   // 0001
         let right = 2;  // 0010
@@ -70,8 +70,8 @@ impl Line2 {
             } else {
                 // failed both tests, so calculate the line segment to clip
                 // from an outside point to an intersection with clip edge
-                let mut x = 0f64;
-                let mut y = 0f64;
+                let mut x = 0f32;
+                let mut y = 0f32;
 
                 // At least one endpoint is outside the clip rectangle; pick it.
                 let outcode_out: OutCode;
@@ -126,7 +126,7 @@ impl Line2 {
 }
 
 impl Index<usize> for Line2 {
-    type Output = Point2<f64>;
+    type Output = Point2<f32>;
     fn index(&self, index: usize) -> &Self::Output {
         &self.points[index]
     }
@@ -142,7 +142,7 @@ impl IndexMut<usize> for Line2 {
 
 pub enum LinePlaneIntersection {
     Line(Line3),
-    Point(Point3<f64>),
+    Point(Point3<f32>),
     None
 }
 
@@ -150,27 +150,27 @@ pub enum LinePlaneIntersection {
 
 /// Represents a 3-dimensional line of finite length. The line is defined by a start point and an end point.
 pub struct Line3{
-    pub points: [Point3<f64>; 2]
+    pub points: [Point3<f32>; 2]
 }
 
 impl Line3{
-    pub fn new(start: Point3<f64>, end: Point3<f64>) -> Line3 {
+    pub fn new(start: Point3<f32>, end: Point3<f32>) -> Line3 {
         Line3 {points: [start, end]}
     }
 
     /// Returns the length of the line.
-    pub fn length(&self) -> f64 {
+    pub fn length(&self) -> f32 {
         (self[1] - self[0]).norm()
     }
 
     /// Returns the direction of the line as a vector.
-    pub fn dir(&self) -> Vector3<f64> {
+    pub fn dir(&self) -> Vector3<f32> {
         self[1] - self[0]
     }
 
     /// Scales the magnitude of the line by a scalar. Both ends of the
     /// line translate.
-    pub fn scale(&self, scale: f64) -> Line3 {
+    pub fn scale(&self, scale: f32) -> Line3 {
         Line3::new(self.points[0] * scale, self.points[1] * scale)
     }
 
@@ -197,9 +197,9 @@ impl Line3{
         LinePlaneIntersection::Point(intersection_point)
     }
 
-    pub fn project(&self, plane: Plane, camera_origin: Point3<f64>) -> Option<Line2> {
+    pub fn project(&self, plane: Plane, camera_origin: Point3<f32>) -> Option<Line2> {
         
-        let mut points: [Point2<f64>; 2] = Default::default();
+        let mut points: [Point2<f32>; 2] = Default::default();
         let visible_line = self.clone();
 
         //Check if the line lies behind the camera
@@ -213,7 +213,7 @@ impl Line3{
             let ray = Ray::new(camera_origin, self.points[i] - camera_origin);
             if let RayPlaneIntersection::Point(projection_3d) = ray.plane_intersection(plane) {
                 let relative_point = projection_3d - plane.origin;
-                points[i] = Point2::<f64>::new(relative_point.dot(&plane.orientation.u), relative_point.dot(&plane.orientation.v));
+                points[i] = Point2::<f32>::new(relative_point.dot(&plane.orientation.u), relative_point.dot(&plane.orientation.v));
             } 
         }        
 
@@ -222,7 +222,7 @@ impl Line3{
 }
 
 impl Index<usize> for Line3 {
-    type Output = Point3<f64>;
+    type Output = Point3<f32>;
     fn index(&self, index: usize) -> &Self::Output {
         &self.points[index]
     }
@@ -237,7 +237,7 @@ impl Rasterize for Line3 {
     fn outline(&self, cam: &Camera) -> Option<Vec<[usize; 2]>>
     {
         if let Some(projected_line) = self.project(Plane::new(cam.orientation, cam.lower_left_corner), cam.origin) {
-            let scale = cam.resoloution.1 as f64/ cam.vertical.norm() as f64;
+            let scale = cam.resoloution.1 as f32/ cam.vertical.norm() as f32;
             if let Some(clipped_line) = projected_line.clip(0.0, cam.horizontal.norm() - 1.0 / scale, 0.0, cam.vertical.norm() - 1.0 / scale) {
                 return Some(clipped_line.scale(scale).bresenham());
             }
@@ -249,18 +249,18 @@ impl Rasterize for Line3 {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use core::f64;
-    use std::f64::consts::PI;
+    use core::f32;
+    use std::f32::consts::PI;
 
 
     #[test]
     fn test_project() {
 
         //Case 1: Axis-aligned camera
-        let v_up = Vector3::<f64>::new(0.0, 1.0, 0.0);
-        let look_from = Point3::<f64>::new(0.0, 0.0, 0.0);
-        let look_at = Point3::<f64>::new(10.0, 0.0, 0.0);
-        let v_fov = 2.0 * 10f64.atan() * 180.0 / PI;
+        let v_up = Vector3::<f32>::new(0.0, 1.0, 0.0);
+        let look_from = Point3::<f32>::new(0.0, 0.0, 0.0);
+        let look_at = Point3::<f32>::new(10.0, 0.0, 0.0);
+        let v_fov = 2.0 * 10f32.atan() * 180.0 / PI;
         let focus_dist = 1.0;
         let aperture = 0.0;
         let aspect_ratio = 1.0;
@@ -269,15 +269,15 @@ mod tests {
         let camera_settings = CameraSettings { look_from, look_at, v_up, v_fov, aspect_ratio, aperture, focus_dist, image_height, image_width };
         let cam = Camera::new(camera_settings);
         
-        let line = Line3::new(Point3::<f64>::new(1.0, 1.0, 1.0), Point3::<f64>::new(1.0, 5.0, 4.0));
+        let line = Line3::new(Point3::<f32>::new(1.0, 1.0, 1.0), Point3::<f32>::new(1.0, 5.0, 4.0));
         let projected_line = line.project(Plane::new(cam.orientation, cam.lower_left_corner), cam.origin).unwrap();
-        assert!((projected_line[0] - Point2::<f64>::new(11.0, 11.0)).norm() < 0.00001);
-        assert!((projected_line[1] - Point2::<f64>::new(14.0, 15.0)).norm() < 0.00001);
+        assert!((projected_line[0] - Point2::<f32>::new(11.0, 11.0)).norm() < 0.00001);
+        assert!((projected_line[1] - Point2::<f32>::new(14.0, 15.0)).norm() < 0.00001);
     }
 
     #[test]
     fn test_bresenham() {
-        let line = Line2::new(Point2::<f64>::new(0.0, 0.0), Point2::<f64>::new(10.0, 0.0));
+        let line = Line2::new(Point2::<f32>::new(0.0, 0.0), Point2::<f32>::new(10.0, 0.0));
         let pixels = line.bresenham();
         
         assert_eq!(pixels.len(), 11);
@@ -290,40 +290,40 @@ mod tests {
     fn test_clip() {
 
         //Case 1: Bottom left corner
-        let start = Point2::<f64>::new(-1.0, -1.0);
-        let end = Point2::<f64>::new(5.0, 5.0);
+        let start = Point2::<f32>::new(-1.0, -1.0);
+        let end = Point2::<f32>::new(5.0, 5.0);
         let line = Line2::new(start, end);
         let clipped_line =line.clip(0.0, 10.0, 0.0, 10.0).unwrap();
-        assert_eq!(clipped_line.points[0], Point2::<f64>::new(0.0, 0.0));
+        assert_eq!(clipped_line.points[0], Point2::<f32>::new(0.0, 0.0));
 
         
         //Case 2: Left-hand side
-        let start = Point2::<f64>::new(-1.0, 5.0);
-        let end = Point2::<f64>::new(5.0, 5.0);
+        let start = Point2::<f32>::new(-1.0, 5.0);
+        let end = Point2::<f32>::new(5.0, 5.0);
         let line = Line2::new(start, end);
         let clipped_line =line.clip(0.0, 10.0, 0.0, 10.0).unwrap();
-        assert_eq!(clipped_line.points[0], Point2::<f64>::new(0.0, 5.0));
+        assert_eq!(clipped_line.points[0], Point2::<f32>::new(0.0, 5.0));
 
                 
         //Case 3: Top
-        let start = Point2::<f64>::new(5.0, 15.0);
-        let end = Point2::<f64>::new(5.0, 5.0);
+        let start = Point2::<f32>::new(5.0, 15.0);
+        let end = Point2::<f32>::new(5.0, 5.0);
         let line = Line2::new(start, end);
         let clipped_line =line.clip(0.0, 10.0, 0.0, 10.0).unwrap();
-        assert_eq!(clipped_line.points[0], Point2::<f64>::new(5.0, 10.0));
+        assert_eq!(clipped_line.points[0], Point2::<f32>::new(5.0, 10.0));
 
         //Case 4: right-hand side
-        let start = Point2::<f64>::new(15.0, 5.0);
-        let end = Point2::<f64>::new(5.0, 5.0);
+        let start = Point2::<f32>::new(15.0, 5.0);
+        let end = Point2::<f32>::new(5.0, 5.0);
         let line = Line2::new(start, end);
         let clipped_line =line.clip(0.0, 10.0, 0.0, 10.0).unwrap();
-        assert_eq!(clipped_line.points[0], Point2::<f64>::new(10.0, 5.0));
+        assert_eq!(clipped_line.points[0], Point2::<f32>::new(10.0, 5.0));
 
         //Case 5: Bottom
-        let start = Point2::<f64>::new(5.0, -5.0);
-        let end = Point2::<f64>::new(5.0, 5.0);
+        let start = Point2::<f32>::new(5.0, -5.0);
+        let end = Point2::<f32>::new(5.0, 5.0);
         let line = Line2::new(start, end);
         let clipped_line =line.clip(0.0, 10.0, 0.0, 10.0).unwrap();
-        assert_eq!(clipped_line.points[0], Point2::<f64>::new(5.0, 0.0));
+        assert_eq!(clipped_line.points[0], Point2::<f32>::new(5.0, 0.0));
     }
 }
